requirements.txt
async-timeout==4.0.2
certifi==2024.7.4
charset-normalizer==2.1.1
evdev==1.9.1
idna==3.7
packaging==21.3
pyaml==25.1.0
pynput==1.7.6
pyparsing==3.0.9
python-dotenv==0.21.0
python-xlib==0.33
PyYAML==6.0.2
redis==4.4.4
requests==2.32.2
six==1.16.0
spotipy==2.25.1
urllib3==1.26.19
